requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
Werkzeug==3.0.1
pyproject==1.3.1
pyproject-toml==0.0.10
psycopg2-binary==2.9.9
pyjwt==2.8.0
argon2-cffi==23.1.0
python-dotenv